Shortened names of tabs + added course explorer tab
I removed instances of "course" from all of the tabs because there was 
overflow that caused the courses selected tab to go below the other 
tabs. There was already this overflow on my small screen before I even 
added the third tab and this commit lets them all sit on line now.