Course explorer (#167)
* Shortened names of tabs + added course explorer tab

I removed instances of "course" from all of the tabs because there was 
overflow that caused the courses selected tab to go below the other 
tabs. There was already this overflow on my small screen before I even 
added the third tab and this commit lets them all sit on line now.

* Populated the explore tab with a new DepartmentList component

Currently, it just displays all the majors as h6s. In the future the 
majors will hopefully be clickable with a drop down of all the courses 
offered in that major.

* Changed from a scroller to a two column layout for the majors

If there are a ton more majors a scroller could still be needed, and if 
thats the case ill add in a dynamic scroller

* Removed scroller in favor of an accordion style

* Merge branch 'master' into course_explorer

* DepartmentList needed data now being passed in as a prop

* Fixed styling so that background shows around button

* even more almost not terrible stylying

* Changed to bootstrap column/rows 

This allows the 2 main columns to be independently sized. Previosuly 
when you clicked a major the box that it is in expanded (rightfully) but 
so did the corresponding box on the other side. This fixes that.

* Maybe not bad styling?

* Fixed multiple buttons having same ID

* beep boop Github Action Master Branch CI Prettier hard at work

* trigger pr deplooy

* Added b-spinner when loading

* let -> const

* Removed uneeded index

* removed empty components and computed

Co-authored-by: Joshua Wu <joshuawu3@gmail.com>
Co-authored-by: I'm a Github Action <beepboop@github.com>