fixed explore page not loading courses on refresh
This seems like a bad fix as the route query I dont think will ever 
return anything. This means on refresh it will use the default semester 
as opposed to the semester that was selected previously. This could 
likely be solved with storing the selected semester in a cookie or 
something along those lines