Credit Cap Feature Added
Adding totalCredits() function to Schedule.js to help calculating credits sum. Adding a if-statement at the end of addCourseSection function to determine whether the credit exceeds credit cap of 21. Throw an error with type "Credit Conflict" when credit is over 21.
Also adding a function notifyCreditConflict() to be in order to have a error message popup when error type "Credit Conflict" is throw and catch.