Update CourseScheduler.vue
I believe the issue occurs because of how the id is set on line 53. When sub semesters occur, they are just set to the same ID, treated as though they are regular semesters. However, since in utils.js the getelementbyID function is used to locate the semester by ID, only one semester can have the ID (according to documentation). So, if both sub semesters are set to the same ID, as I suspect they are, then the function can not properly access the sub semester selected. Attempted to try and change where the id was set but did not compile - will continue working on that.