Image export bug (#447)
* Update CourseScheduler.vue

Issue not fixed yet, attempt to modify the string that is passed to the export image function. Thinking the issue relates to the sub semesters not having an actual course name as Fall 2021 etc do.

* Update CourseScheduler.vue

I believe the issue occurs because of how the id is set on line 53. When sub semesters occur, they are just set to the same ID, treated as though they are regular semesters. However, since in utils.js the getelementbyID function is used to locate the semester by ID, only one semester can have the ID (according to documentation). So, if both sub semesters are set to the same ID, as I suspect they are, then the function can not properly access the sub semester selected. Attempted to try and change where the id was set but did not compile - will continue working on that.

* Update CourseScheduler.vue

image export of sub semesters now works

* Update utils.js

update for image export issue

* Update utils.js

cleanup

* Update CourseScheduler.vue

cleanup

* Update CourseScheduler.vue

cleanup

* Update CourseScheduler.vue

Needed to update due to implementation of Color Blind Assistance