Adding to new Features screen
Edited 2 files to help fix the blocker that I ran into the other day.
This will bring us closer to arriving at a screen with all majors of the school.
Next: Will add the templates for each major using the json file with a map of all majors and its given template.