Overhaul of Prerequisites.vue
I overhauled the entire page to make it simpler for me to understand. I made it so that you have 2 input text boxes, one for department code and one for course number, this will output the course onto the page. I then created an erase button in case you want to get rid of a specific course, then I made a function that will delete it off the site with a click on it.