Dark mode
Minor changes so that button changes color for a moment to signal the user that the button got clicked, however, for the grid box, the color change to the button is not permanent, so the user is not able to tell which buttons got clicked and which weren't. All of them appear to be dark mode color.