Change Font Size to be More Responsive
The font size now changes based on the width of the browser window, which prevents the day labels from taking two lines and as a result prevents some columns from shifting down. However, when the browser window becomes smaller, the text becomes hard to read when it has the space to be a bigger font size. A possible solution is to create a new function that would change the font size based on the width of the grid boxes.