implemented cross template wildcards
now, if there are characters after the wildcard (*), it will treat the wildcard as a variable that must hold a consistent value across all templates that use the same token.

For example, if we use concentration.*1 on two different templates, then that wildcard must resolve to the same concentration for both templates.

Using different tokens, like concentration.*1 and concentration.*2, or using the default concentration.*, allows the templates to evaluate to different wildcards for each template (like the way it was before).