schedule switching and wildcard selection fully functional
user input will now return a dictionary of variables to update within vue.

but now I think about it, the only component using this is the schedule where the schedule name is returned when user switches schedule. However, schedules should not be handled in the frontend in the first place...