Create toggle to switch between standard and military time
I created the toggle to switch between the two time representations. The plan is to have the website start out already toggled on standard time or toggled off for military time. Functions in the Schedule page and other elements that display time will call toggleTime() and calculate based on which representation is used. Another way is to calculate for both representations and choose which to display depending on the returned result from toggleTime().