brought back sorting recommendations
recommendation paths due to wildcards now sort by the number of courses already fulfilled in that category.

This is done by adding a sort function to my dict_array class, and it works by sorting the array elements and allows the calling of a function within the array for a custom sort.

For example, if the array elements are dictionaries, we can sort by their dictionary entry of "value" by calling dict_array.sort_elements('get', 'value') which gets converted into element.get('value') to fetch the key for sorting.