UI overhaul (๑>ᴗ<๑)
Instead of a card for each requirement, I made a new card style that combines similar requirements together. Requirements are grouped together based on the names.

So, if the name contains a dash, like "core - data structures", the part in front of the first dash is considered the group. If there's no dash, the entire name becomes the group.

So you can have things like "core - DS", "core - FOCS", "math - calc 1", "math - calc 2" and there will only be two cards, one for core, one for math. Nice and tidy~