support for direct evaluation with @ symbol
adding @ symbol in front of a token will evaluate the value itself rather than the truth. So if we want to make sure two wildcard resolutions are not the same, we can use @concentration.*1 != @concentration.*2. Without the @ symbol, we would be instead evaluating the presence of each of those concentrations in the course and then comparing those booleans.