support for group oasis
basically a group of requirements that are considered separately from other requirements. Useful when you want the group to be non replaceable internally but replaceable globally.

Ideally we should use a graph to dictate which groups are not allowed to share courses with each other, but that would either require a subpar algorithm or require only cliques/clique with tails.