Degree Planner checkpoint - time for SQL?
So I will need to start working on data retention methods like setting up a SQL table, but for now I think the API/frontend is in a usable state, more or less.

There are still plenty of web features I plan on adding before an official release, and (if my sanity allows) I could rewrite the fulfillment algorithm with a more robust method for better generalization with various degree templates... but that's a problem for future me, right?