Updated hour labels for schedule grid and finals page
The hour labels for the schedule grid on the schedule and finals page are updated but I'm unable to check because of file issues. The finals info on the finals page should also be updated to be similar to the course info on the schedule page. However, I'm unable to test it because the finals page isn't working (can't see any finals).