added setdegree api call
newuser api call on longer handles setting degree/schedule info, use their own api calls for that.

However, newuser will return a "basic info" dictionary, which as of now, returns a directory tree containing degrees.

This directory tree is a new data structure implemented into the Attributes class, which can be used for a variety of user data handling like organizing schedules (although is overkill for that as of now). But it's still nice to have for whatever.