723 more detailed explore page with sorting options (#753)
* Added buttons and started implementing functions

The buttons to list the departments by alphabetical order and by school were created. The functions to sort the departments alphabetically and separated into two columns are beginning to be implemented.

* Added functions to sort departments alphabetically

The functions alphDict() and alphDepartmentObjects() were implemented to help alphabetically sort the departments. However, I was unable to test it out because the website keeps saying Loading Departments.

* Fixed code so that it appears on the website

The code is fixed so that the alphabetical sections with the respective departments show on the website. However, each alphabetical section are not in alphabetical order.

* Fix how alphabetical sections were arranged

The departments are now sorted by their first letter and each section is separated into two columns with the first 13 letters in the first column and the rest in the second column.

* Fixed so that website don't display empty sections

Previously, the website had a section for every letter in the alphabet even if there were no departments that start with that letter. Now, it only displays the sections that are not empty (has departments that start with the letter).

* Remove extra code

Removed alphabet array in alphDeptDict() because it wasn't being used.

* Found where department code long names are stored

The long names for the department codes INQR and STSO are added, but it still doesn't display on the website.

* HASS typo

fixed HASS typo

---------

Co-authored-by: jonha1 <jonnyha888@gmail.com>