718 finals schedule days don't line up (#737)
* Change Font Size to be More Responsive

The font size now changes based on the width of the browser window, which prevents the day labels from taking two lines and as a result prevents some columns from shifting down. However, when the browser window becomes smaller, the text becomes hard to read when it has the space to be a bigger font size. A possible solution is to create a new function that would change the font size based on the width of the grid boxes.

* Try to Implement Function to Change Font Size Based on Grid Box Size

The function is supposed to take in the length of day label and if it's longer than the length of the grid box, then it will shrink the font size. However, it is currently not working. I made the font size slightly bigger compared to last time.

* Revert back to just using vw and not a function

Function still doesn't work, so code is reverted back to just using vw. Font size is updated to 0.85 vw instead of 0.9vw because the grid lines don't line up at the larger size.