Proof of Work rate limiting
Users need a userid to query calculations, which is used to both rate limit and to retrieve answers (wait actually there's literally no reason not to combine these two into one wtf I'll fix that later)

anyways we still need a userid for rate limiting I guess, so all is good an my efforts were not wasted.

Anyways, the way it works is the server sends a "hash head" to the client. The client must then append whatever they want to this head, so long as the resulting hash (SHA256) contains 3 (configurable amount) leading zeroes in hex representation.

This basically forces the client to brute force an ID that works. A table won't work since they must use the hash head given to them.

3 leading zeroes isn't a big requirement, it's done in a few hundred milliseconds, mostly compatible with mobile and will alleviate simple spam attacks to an extent.

The whole proof of work thing is meant for accountless interactions, so if accounts are to be required to use degree planner this is no longer necessary. However, for now I think accountless should be kept available in all but the highest traffic situations.